pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zcyber-xhs"
version = "0.1.0"
description = "Automated cybersecurity content pipeline for Xiaohongshu"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "httpx>=0.27",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "click>=8.1",
    "apscheduler>=3.10",
    "python-dotenv>=1.0",
    "playwright>=1.40",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
zcyber = "zcyber_xhs.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/zcyber_xhs"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
